requirements.txt
flask
langchain
langchain-community
langchain-core
openai
python-dotenv
